pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pem_utilities"
version = "1.0.0"
description = "Comprehensive electromagnetic simulation utilities for antenna modeling, beamforming, materials, and visualization"
readme = "readme.md"
license = {text = "MIT"}
authors = [
    {name = "ANSYS", email = "support@ansys.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
keywords = ["electromagnetic", "antenna", "beamforming", "simulation", "radar", "communications"]
requires-python = ">=3.10,<3.13"

dependencies = [
    "numpy==1.26.4",
    "scipy",
    "matplotlib",
    "Pillow==10.4.0",
    "opencv-python",
    "vtk==9.3.0",
    "pyvista==0.44.1",
    "imageio-ffmpeg",
    "imageio",
    "tqdm",
    "netCDF4",
    "scikit-image",
    "pyargus",
    "pycollada",
    "ipykernel",
    "mitsuba",
    "utm",
    "python-srtm",
    "SRTM.py",
    "osmnx",
    "usd-core",
    "protobuf",
    "geopandas",
    "rasterio",
    "h5py",
    "PySide6",
    "pyvistaqt",
    "pandas",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]
mediapipe = [
    "mediapipe",  # for web-cam based radar demo
]

[project.urls]
Homepage = "https://github.com/ansys/perceive_em"
Documentation = "https://github.com/ansys/perceive_em"
Repository = "https://github.com/ansys/perceive_em"
"Bug Tracker" = "https://github.com/ansys/perceive_em/issues"

[tool.setuptools]
packages = ["pem_utilities", "pem_toolkits"]
include-package-data = true

[tool.setuptools.package-data]
"pem_utilities" = ["*.json", "*.ffd"]
"pem_toolkits" = ["*.json", "*.ffd"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"